Added locations north and northwest of town
Took time to understand the basics of RPG Maker MV, and made some of the locations above the town, as well as allow the player to access these sections. Feel free to look over or correct my work if necessary. Chests need to be added to the NW room (the maze), and weather effects will need to be added.